refactor: renamed EventResponse into EventEntity
